# New: Exception message template for uncaught rule error (refs #5221)