--- a/lib/eslint.js
+++ b/lib/eslint.js
@@ -27,7 +27,8 @@ var lodash = require("lodash"),
     validator = require("./config/config-validator"),
     replacements = require("../conf/replacements.json"),
     assert = require("assert"),
-    CodePathAnalyzer = require("./code-path-analysis/code-path-analyzer");
+    CodePathAnalyzer = require("./code-path-analysis/code-path-analyzer"),
+    ESLINT_VERSION = require("../package.json").version;
 
 var DEFAULT_PARSER = require("../conf/eslint.json").parser;
 
@@ -505,6 +506,52 @@ function stripUnicodeBOM(text) {
     return text;
 }
 
+/**
+ * Wraps a rule action with an exception handler.
+ *
+ * The exception handler will catch exceptions thrown by rule actions, decorate
+ * the exception with rule metadata, and rethrow the exception. This will allow
+ * us to detect the uncaught exception as a rule crash and show a user-friendly
+ * error message.
+ *
+ * @param {function} action The rule action to wrap.
+ * @param {object} metadata Metadata that are useful for bug reports.
+ * @returns {function} A function which will catch, decorate, and rethrow rule exceptions.
+ */
+function createRuleExceptionHandler(action, metadata) {
+    var ruleName = metadata.ruleName,
+        fileName = metadata.fileName,
+        parserPath = metadata.parserPath,
+        hasDefaultParser = parserPath === DEFAULT_PARSER,
+        pluginDelimiterIndex = ruleName.indexOf("/"),
+        pluginName = null,
+        eightSpaces = new Array(9).join(" ");
+
+    if (pluginDelimiterIndex > -1) {
+        pluginName = ruleName.slice(0, ruleName.indexOf("/"));
+    }
+
+    return function() {
+        try {
+            action.apply(null, arguments);
+        } catch (err) {
+            err.messageTemplate = "rule-crashed";
+            err.messageData = {
+                ruleName: ruleName,
+                pluginName: pluginName,
+                fileName: fileName,
+                parserPath: parserPath,
+                hasDefaultParser: hasDefaultParser,
+                eslintVersion: ESLINT_VERSION,
+                errorMessage: err.message,
+                stackTrace: err.stack.replace(/^\s+/gm, eightSpaces)
+            };
+
+            throw err;
+        }
+    };
+}
+
 //------------------------------------------------------------------------------
 // Public Interface
 //------------------------------------------------------------------------------
@@ -773,9 +820,16 @@ module.exports = (function() {
 
                     // add all the node types as listeners
                     Object.keys(rule).forEach(function(nodeType) {
+                        var action = rule[nodeType],
+                            wrappedAction = createRuleExceptionHandler(action, {
+                                ruleName: key,
+                                fileName: currentFilename,
+                                parserPath: config.parser
+                            });
+
                         api.on(nodeType, timing.enabled
-                            ? timing.time(key, rule[nodeType])
-                            : rule[nodeType]
+                            ? timing.time(key, wrappedAction)
+                            : wrappedAction
                         );
                     });
                 } catch (ex) {