Hi folks. I'm asking for [30% feedback](http://blog.sandglaz.com/30-percent-feedback-rule/) here (meaning this PR is probably not ready to merge yet). Please take a look and let me know if this is the right direction.

Terminology:

* Rule action: That is my name for the node event handler in rules, i.e., the values returned in the rule creator function.

The highlights:

* When lib/eslint.js calls `api.on(nodeType)`, I am wrapping the rule action function in a containing function which will call the rule action with try/catch, and if an exception is thrown, it will decorate the exception with metadata representing a new message template.
* New message template contains information on how to report the bug. Probably needs some polishing and copy-editing.

Done:

- [x] Get the filename into the error message
- [x] Mention plugin version as an important piece of information for a plugin bug
- [x] Show parser name, distinguish between default and non-default parser

Do we want these for an initial PR?
- [ ] Different top-level message for default vs non-default parser?
- [ ] Report on node metadata (e.g., line and column number)?
- [ ] Crazy thought: Get the RuleContext object into the wrapper, so that we can do `context.getSourceCode().getText(node)`?

Feedback welcome!