@platinumazure 2016-03-07
I aim to do some incremental improvements tonight (US time):

* Get filename into error message
* Mention plugin version as an important piece of information for a bug