@ilyavolodin 2016-03-07
@platinumazure `ruleContext` has filename in it. So you should be able to extract it from there. There is also a reference to the `parser` attached to `ruleContext`, but I think it's the whole executable object, and not just the name (although I might be wrong).