@platinumazure 2016-03-07
Thanks @ilyavolodin for the tips. I wasn't too worried about the filename since I knew `RuleContext` was constructed around a parsed file, but thanks for saving me a few minutes of digging. And I'll check out the parser as well. (I thought I remembered a `parserName` property floating around somewhere, but that might not be on the `RuleContext` instance.)