@ilyavolodin 2016-03-07
Actually just checked. `parserName` should also be available on `ruleContext`. It was added in #4649