@BYK 2016-03-07
Great idea btw, thanks for the PR!