@platinumazure 2016-03-08
Updated with file name, parser name (and whether it represents the default parser or not), and a note about installed plugin version being important.

I've modified my initial post to cover what's done and what's not yet done. I do think it might be worthwhile to limit the scope of this PR to focus on high-value low-hanging fruit, and we can then focus on some of the other enhancements later.