@platinumazure 2016-03-08
At the request of @ilyavolodin, here are the perf metrics:

* Single-file: 12784.163798ms to 14395.568591ms (12.6% slower)
* Multi-file: 40937.154772ms to 46614.713000ms (13.9% slower)

Yeah, not great.

Since I'm allowing the exception to propagate anyway, there has to be a way for me to put a wrapper around the estraverse traversal instead of the rule actions. I will look into that.

Full dumps:

Before:

> > eslint@2.3.0 perf C:\Users\kpartington\Documents\GitHub\eslint
> > node Makefile.js perf
> 
> 
> Loading:
>   Load performance Run #1:  852.383636ms
>   Load performance Run #2:  861.442609ms
>   Load performance Run #3:  847.120133ms
>   Load performance Run #4:  860.691636ms
>   Load performance Run #5:  867.616382ms
> 
>   Load Performance median:  860.691636ms
> 
> 
> Single File:
>   CPU Speed is 2294 with multiplier 13000000
>   Performance Run #1:  12920.209023ms
>   Performance Run #2:  12696.805085ms
>   Performance Run #3:  12798.272539ms
>   Performance Run #4:  12782.383417000001ms
>   Performance Run #5:  12784.163798ms
> 
>   Performance budget exceeded: 12784.163798ms (limit: 5666.957279860506ms)
> 
> 
> Multi Files (450 files):
>   CPU Speed is 2294 with multiplier 39000000
>   Performance Run #1:  40914.100396ms
>   Performance Run #2:  40983.377754ms
>   Performance Run #3:  40652.639136ms
>   Performance Run #4:  40937.154772ms
>   Performance Run #5:  41096.501573ms
> 
>   Performance budget exceeded: 40937.154772ms (limit: 17000.871839581516ms)

After:

> > eslint@2.3.0 perf C:\Users\kpartington\Documents\GitHub\eslint
> > node Makefile.js perf
> 
> 
> Loading:
>   Load performance Run #1:  882.735344ms
>   Load performance Run #2:  888.435241ms
>   Load performance Run #3:  894.15611ms
>   Load performance Run #4:  892.627838ms
>   Load performance Run #5:  889.052351ms
> 
>   Load Performance median:  889.052351ms
> 
> 
> Single File:
>   CPU Speed is 2294 with multiplier 13000000
>   Performance Run #1:  14395.568591ms
>   Performance Run #2:  13943.606535ms
>   Performance Run #3:  14556.950114ms
>   Performance Run #4:  14166.329555ms
>   Performance Run #5:  15868.433394ms
> 
>   Performance budget exceeded: 14395.568591ms (limit: 5666.957279860506ms)
> 
> 
> Multi Files (450 files):
>   CPU Speed is 2294 with multiplier 39000000
>   Performance Run #1:  49652.101899ms
>   Performance Run #2:  46757.982824ms
>   Performance Run #3:  46517.312312ms
>   Performance Run #4:  46065.083868ms
>   Performance Run #5:  46614.713ms
> 
>   Performance budget exceeded: 46614.713ms (limit: 17000.871839581516ms)