@platinumazure 2016-03-08
Okay, I've done a quick rewrite to surround `controller.traverse()` with a try/catch.

Perf results:
* Single file: 12844.976982ms (0.5% slower)
* Multi file: 41414.348267ms (1.2% slower)

...And I've already run into the first hurdle. I also lose the name of the rule that threw by going to a more higher-level handler. (I could try parsing it from the error's stack trace, but...) This also means we can't easily tell if the rule was a core rule or plugin rule.

@ilyavolodin What do you think? Do we accept the loss of the rule name and plugin and trust that users (or we) can read a stack trace, or do we accept the performance hit from the original approach and I need to look for ways to offset it?