@ilyavolodin 2016-03-08
@platinumazure I'm trying to think if this is a right approach in general. Taking perf hit of 1.2% (and it's going to be larger as the number of files increases) for error tracking might be a bit pushing it. Maybe instead we should try to use global error handler and parse out as much information as we can from stack trace? I know we are not going to get nearly as much information, and also it will not be easy to parse stack out, but in that case we will only get hit when the error occurs vs. every time. Not really sure, honestly. @eslint/eslint-team what do you think?