@btmills 2016-03-08
What if we had an error helper singleton that allowed us to tell it whenever we start lifting another file the name of the file and the config we're using? If something bad happens, we'd have the stack trace along with the most recent source file and config.