@ilyavolodin 2016-03-08
I think @btmills suggestion might work. If we would have singleton available that we could populate with the information about the last rule that was executing, then we could use global error handler to catch errors and still retrieve enough information for message to be useful.