@alberto 2016-03-08
The only catch is this could be a problem if we want to use multiple processes in the future.