@platinumazure 2016-03-08
@ilyavolodin @alberto That doesn't really have much to do with this PR. It
is true that multiple processes means it will be harder to report errors.
But we would need to aggregate success results in a multiprocess
environment, and we should be able to aggregate errors at that time. From
my perspective, there is already a redesign required once we go parallel. I
don't think that should affect this issue unless we know that
parallelization is about to happen (in which case this single-process
implementation is a waste of time).
On Mar 8, 2016 9:04 AM, "Ilya Volodin" <notifications@github.com> wrote:

> Ahh.. very good point! Didn't think about that. Hmm... that might be a
> problem.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/eslint/eslint/pull/5483#issuecomment-193816047>.
>