@ilyavolodin 2016-03-08
@platinumazure Multiple processes discussion was about global error handler with singleton info collector.