@nzakas 2016-03-08
I'm not sure the perf hit is worth it, either. I'd rather start with areas where we are already throwing errors, as we can just annotate those, and cycle back to things like this. I think we also need a strategy for what we output to the user vs. what we write to a log file. The example in this PR seems more like log file information.

In short, I think we need more discussion around how we want to proceed with better error messaging.