@platinumazure 2016-03-08
Closing, since it seems the PR was premature. (By the way, having the issue be "accepted" sends a different message than that.)